app/utils/token_tracker.py
from sqlalchemy.ext.asyncio import AsyncSession
from redis import asyncio as aioredis
from datetime import datetime
import logging

logger = logging.getLogger(__name__)

class TokenTracker:
    def __init__(self, user_id: int, model: str, db: AsyncSession, redis: aioredis.Redis):
        self.user_id = user_id
        self.model = model
        self.db = db
        self.redis = redis
    
    async def track_usage(self, input_tokens: int, output_tokens: int, request_id: str):
        """
        Track token usage and update user's balance
        """
        try:
            # For now, just log the usage
            logger.info(f"User {self.user_id} used {input_tokens} input and {output_tokens} output tokens")
            
        except Exception as e:
            logger.error(f"Error tracking token usage: {str(e)}")
